moving common/simple to common
