+++ clusterloader2/pkg/measurement/common/resource_usage.go
@@ -14,7 +14,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package simple
+package common
 
 import (
 	"fmt"
